Move `inherits_doc_hidden` and `should_ignore_res` into `clean/utils.rs`
